RD-6765 Move creating system filters to configure_manager
With this, system filters are always kept up-to-date, both on the
regular c-m-i AIO install, and in the split-services install as well.

Thankfully, the existing function already checks what exists, and
only inserts the missing pieces.

Still leave the original script around, because things like upgrade
or snap-restore use it.
